Refactor code structure into a package and provide single entry point, along with precached api responses
