+++ .gitignore
@@ -11,6 +11,7 @@ url_sample_images/
 
 # Ignore generated collages
 collages/
+old_collages/
 
 # Ignore common output and cache files (in case)
 *.pyc
