+++ collage_server/image_embedder.py
