+++ run_collage_server.py
@@ -1,0 +1,4 @@
+from collage_server.server import app
+
+if __name__ == '__main__':
+    app.run(debug=True)
